+++ RelationalAI.Examples/CancelTransaction.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using System.Collections.Generic;
+using System.CommandLine;
+using System.CommandLine.NamingConventionBinder;
+
+namespace RelationalAI.Examples
+{
+    public class CancelTransaction
+    {
+        public static Command GetCommand()
+        {
+            var cmd = new Command("CancelTransaction", "--id <Transaction id>"){
+                new Option<string>("--id"){
+                    IsRequired = true,
+                    Description = "Transaction id."
+                },
+
+                new Option<string>("--profile"){
+                    IsRequired = false,
+                    Description = "Profile name from .rai/config to connect to RAI Cloud."
+                }
+            };
+            cmd.Description = "Cancel an asynchronous transaction.";
+            cmd.Handler = CommandHandler.Create<string, string>(Run);
+            return cmd;
+        }
+
+        private static void Run(string id, string profile = "default")
+        {
+            Dictionary<string, object> config = Config.Read("", profile);
+            Client.Context context = new Client.Context(config);
+            Client client = new Client(context);
+            var rsp = client.CancelTransaction(id);
+            Console.WriteLine(rsp);
+        }
+    }
+}